Change developer options strings to something less dumb
In the past we gone ahead and just enabled developer options by default and
just removed the code used for 'tap to enable'. This was never an issue but
considering some options like adaptive options are now booby trapped by Google
with developer options, it doesn't make sense to do that in O and maybe future
versions. (see Pixel launcher and Launcher3)

With this said, we need better strings. Users need to know that messing with
developer options is not a joke and could end up slowing their device down or
making it unusable for them if they're too noobish.

Change-Id: Ie34ab8df5d6cea496099d0f1c4eda19c08d164b4